Style Guide Update; Changed to GLES 2
Made a lot of changes to all the script files to make them better
conform to the current GDScript style guide recommendations as
well as general readability updates. Attempted to add stronger
typing in a few places, and reordered the api request function
to be less verbose in the most cases (i.e. parameters and
session_id were swapped). Everything still functions as
intended as far as I can tell.

Also changed the renderer for the project to GLES 2 because the
HTML5 versions of games can't really take advantage of GLES 3
and for highest usability developers should stick to GLES 2 for
web games anyway.

Things to note:
- Classes for components should probably be moved to their own
script files and I may do that myself if I have a chance to do
so.
- Attempted to add debug as an export variable to the API since
there is no chance to call the API request anywhere in the
original code with the debug parameter. This would theoretically
make it possible for the end user developer to simply click a
checkbox on or off for testing/release, but the debug parameter
doesn't seem to work properly (at least according to how the
NGio docs say it should).
- Separated script of the Sample scene to it's own file. It might
even be most desirable to make separate scripts for each button
so that each function can be viewed in isolation, right now the
sample code is kind of a mess to try and read.